acpi_hpet: correctly get number of timers/comparators in a timer block
Also, account for a quirk of AMD/ATI HPET which reports number of timers
instead of id of the last timer as manadated by the specification.
Currently this has no effect on functionality but in the future we may
make actual use of the HPET timers, not only of its timecounter.

MFC after:	2 weeks


git-svn-id: http://svn.freebsd.org/base/head@203062 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f