sh: Fix various things about expansions: * remove the backslash from \} inside double quotes inside +-=?   substitutions, e.g. "${$+\}a}" * maintain separate double-quote state for ${v#...} and ${v%...};   single and double quotes are special inside, even in a double-quoted   string or here document * keep track of correct order of substitutions and arithmetic
This is different from dash's approach, which does not track individual
double quotes in the parser, trying to fix this up during expansion.
This treats single quotes inside "${v#...}" incorrectly, however.

This is similar to NetBSD's approach (as submitted in PR bin/57554), but
recognizes the difference between +-=? and #% substitutions hinted at in
POSIX and is more refined for arithmetic expansion and here documents.

PR:		bin/57554
Exp-run done by:	erwin (with some other sh(1) changes)


git-svn-id: http://svn.freebsd.org/base/head@206145 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f