indent(1): don't treat bare '_t' as a type name when -ta is used
It seems that identifier "_t" is sometimes used as a variable name,
even in our tree.  Not that I endorse that, but still it's better
to require at least one character before _t suffix to consider
an identifier to be a type name.

Reported by:	Alex Vasylenko <lxv@omut.org>
MFC after:	1 week


git-svn-id: http://svn.freebsd.org/base/head@206687 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f