Revert r209544. It papers over a binutils bug, and is not even a reliable fix for it.
The bug occurs when using the --as-needed flag to ld in the presence of
synthetic linker-generated symbols that reference symbols defined in
linked-to shared libraries with versioned symbols. When the only symbols
used from a library fall into this category, ld will drop the DT_NEEDED
entry for it, but retain the versioning information. This bug is best
fixed/hacked around in binutils, not in rtld.

Discussed with:	kan


git-svn-id: http://svn.freebsd.org/base/head@209558 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f