Since r186119 IP6 input counters for octets and packets were not working anymore.  In addition more checks and operations were missing.
In case lla_lookup results in a match, get the ifaddr to update the
statistics counters, and check that the address is neither tentative,
duplicate or otherwise invalid before accepting the packet.  If ok,
record the address information in the mbuf.  [ as is done in case
lla_lookup does not return a result and we go through the FIB ].

Reported by:	remko
Tested by:	remko
MFC after:	2 weeks


git-svn-id: http://svn.freebsd.org/base/head@210350 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f